requirements.txt
psycopg2-binary==2.9.3
Faker==13.11.1
Django==4.0.4
django-environ==0.8.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
environ==1.0
django-filter==21.1
Pillow==9.1.0
